requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
djangorestframework==3.9.2
idna==2.8
oppai==3.2.1
oppaipy==1.0.2
psycopg2==2.8.1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.1
